requirements.txt
RPi.GPIO
pynmea2
pyserial
smbus2
adafruit-blinka
adafruit-circuitpython-vl53l0x
adafruit-circuitpython-busdevice
opencv-python